@@ -18,10 +18,13 @@ namespace DBHelper
 				var key = string.Concat("@", p.ParameterName);
 				if (value == null)
 					sql = GetNullSql(sql, key);
-				else if (Regex.IsMatch(value, @"(^(\-|\+)?\d+(\.\d+)?$)|(^SELECT\s.+\sFROM\s)|(true)|(false)",
-					RegexOptions.IgnoreCase) && !isString.Contains(p.NpgsqlDbType)) sql = sql.Replace(key, value);
+
+				else if (Regex.IsMatch(value, @"(^(\-|\+)?\d+(\.\d+)?$)|(^SELECT\s.+\sFROM\s)|(true)|(false)", RegexOptions.IgnoreCase) && !isString.Contains(p.NpgsqlDbType))
+					sql = sql.Replace(key, value);
+
 				else if (value.Contains("array"))
 					sql = sql.Replace(key, value);
+
 				else
 					sql = sql.Replace(key, $"'{value}'");
 			}
@@ -43,7 +46,7 @@ namespace DBHelper
 			Type type = value.GetType();
 			if (type.IsArray)
 			{
-				var arrStr = (value as object[]).Select(a => $"'{a.ToEmptyOrString()}'");
+				var arrStr = (value as object[]).Select(a => $"'{a?.ToString() ?? ""}'");
 				return $"array[{string.Join(",", arrStr)}]";
 			}
 			return value?.ToString();
