@@ -6,7 +6,7 @@ public class HostConfig
 	/// <summary>
 	/// 默认数据库配置
 	/// </summary>
-	public const DatabaseType DefaultDatabase = DatabaseType.Master;
+	public const DatabaseType DEFAULT_DATABASE = DatabaseType.Master;
 	/// <summary>
 	/// 全局配置
 	/// </summary>
