@@ -9,6 +9,7 @@ using System;
 using DBHelper;
 using System.Text;
 using CodeFactory;
+using System.Linq;
 
 namespace dotnetCore_pgsql
 {
@@ -21,7 +22,7 @@ namespace dotnetCore_pgsql
 			Console.InputEncoding = Encoding.GetEncoding("UTF-8");
 			Console.WriteLine(@"
 ##########################################################
-#     .net core 2.1+npsql 4.0.2 Postgresql Code Maker    #
+#     .net core 3.0+npsql 4.1.1 Postgresql Code Maker    #
 #                author by leisaupei                     #
 #      https://github.com/leisaupei/dotnetCore_pgsql     #
 ##########################################################
@@ -33,13 +34,26 @@ namespace dotnetCore_pgsql
   - db		database name
   - path	output path
   - name	project name
-  - type    database enum type name *not required
-> Example: host=localhost;port=5432;user=postgres;pwd=123456;db=postgres;name=test;path=d:\workspace;type=master
+  - type    database enum type name, 'master' if only one. *not required
+> Example: host=localhost;port=5432;user=postgres;pwd=123456;db=postgres;name=test;path=d:\workspace\test;type=master
+
+> Multiple Example, slipt from ',': 
+	host=localhost;port=5432;user=postgres;pwd=123456;db=postgres;name=test;path=d:\workspace\test;type=master,host=localhost;port=5432;user=postgres;pwd=123456;db=postgres;name=test;path=d:\workspace\test;type=xxx
 ");
-			if (!args.IsNullOrEmpty())
-				LetsGo.Produce(args[0]);
+			if (args?.Length > 0)
+			{
+				if (args[0].Contains(","))
+				{
+					foreach (var item in args[0].Split(","))
+						LetsGo.Produce(item);
+				}
+				else
+					LetsGo.Produce(args[0]);
+			}
 			else
+			{
 				LetsGo.Produce(Console.ReadLine());
+			}
 			Console.WriteLine("successful...");
 			Console.ReadLine();
 		}
